playlists, dice assign, fix bug unicode detect

+++ events/messageCreate.js
@@ -17,7 +17,7 @@ module.exports = {
         }
 
         //check if message has an unicode character
-        if (message.content.match(/[^\x00-\x7F]/g)) {
+        if (message.content.match(/[^\x00-\x7F]/g) && !message.content.match(/<a?:.+?:\d{18}>|\p{Extended_Pictographic}/gu)) {
             const m = await message.reply('Message contains an unicode character')
             message.delete()
             setTimeout(function () { m.delete() }, 5000)
